requirements.txt
groq
requests
python-dotenv
feedparser
pydantic
streamlit
